+++ lib/main.dart
@@ -1,28 +1,20 @@
-import 'package:deepdive_application/pages/intro_page.dart';
 import 'package:flutter/material.dart';
+import 'pages/intro_page.dart';
 
 void main() {
   runApp(const MyApp());
 }
 
-class MyApp extends StatefulWidget {
+class MyApp extends StatelessWidget {
   const MyApp({super.key});
 
-  @override
-  State<MyApp> createState() => _MyAppState();
-}
-
-class _MyAppState extends State<MyApp> {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       theme: ThemeData(
-        textSelectionTheme: TextSelectionThemeData(
-          cursorColor: Color(0xFF338BEF), // 커서 색상을 파란색으로 설정
-        ),
+        fontFamily: 'Pretendard', //지정폰트 세팅 : 에셋에 폰트 생성, yaml파일에도 활성화
       ),
-      // home: IntroPage(),
-      // 나머지 앱 설정
+      home: IntroPage(),
     );
   }
 }
