--- lib/pages/registration/primary_button.dart
@@ -1,43 +1,0 @@
-import 'package:flutter/material.dart';
-
-class PrimaryButton extends StatelessWidget {
-  final String text; // 버튼에 표시될 텍스트
-  final VoidCallback onPressed; // 버튼 클릭 시 실행할 함수
-
-  const PrimaryButton({
-    Key? key,
-    required this.text,
-    required this.onPressed,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      margin: const EdgeInsets.only(bottom: 0),
-      child: Padding(
-        padding: const EdgeInsets.symmetric(
-          horizontal: 16,
-          vertical: 12,
-        ),
-        child: ElevatedButton(
-          onPressed: onPressed,
-          style: ElevatedButton.styleFrom(
-            backgroundColor: Color(0xFF0770E9),
-            minimumSize: Size(double.infinity, 50),
-            shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(4),
-            ),
-          ),
-          child: Text(
-            text, // 전달받은 텍스트 사용
-            style: TextStyle(
-              fontSize: 18,
-              fontWeight: FontWeight.bold,
-              color: Colors.white,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
